Try catch around user supplied function
Promises are supported inside of uncaughtException, SIGINT and SIGTERM but not exit. Most restarts call SIGINT or SIGTERM before they exit.

I don't know why you removed the check for an infinite loop of uncaught exceptions? If there is an uncaught exception inside of the user defined function or our function it will loop forever. I added a try catch block around the user function so it won't happen for user functions but if you don't want a check for it we need to be 100% sure our code won't cause a loop