Moved cache patching and added 15s timeout for resuming
Moved cache patching to WebSocketManager and added timeout of 15s to remove the loadCache event. Now only attaches the event if a session ID is present to stop the event being attached on an identify connection attempt

I don't know the best way to remove the event listener in the timeout so let me know how you would do it.